+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Standartized return type of loadUserPhoto and loadUserNamePronunciation [#4402](https://github.com/rokwire/illinois-app/issues/4402).
 - Cleaned up social message deep link support [#4572](https://github.com/rokwire/illinois-app/issues/4572).
 ### Added
+- Add support for editing messages on the social block [#529](https://github.com/rokwire/app-flutter-plugin/issues/529)
 - Add deep links for messages [#516](https://github.com/rokwire/app-flutter-plugin/issues/516)
 - Use Social BB for Posts, Comments and Reactions (task in progress) [#498](https://github.com/rokwire/app-flutter-plugin/issues/498).
 - Acknowledge event.notification_settings field and APIs [#4478](https://github.com/rokwire/illinois-app/issues/4478).